Move PackageGenerator and PackageListManager to their own structs
