+++ Sources/DangerDependenciesResolver/Data+Encoding.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+extension Data {
+    func decoded<T: Decodable>() throws -> T {
+        let decoder = JSONDecoder()
+        return try decoder.decode(T.self, from: self)
+    }
+}
+
+extension Encodable {
+    func encoded() throws -> Data {
+        let encoder = JSONEncoder()
+        return try encoder.encode(self)
+    }
+}
+
+
